WIP permissions
* This state sets a low tested env of permissions working.

* It's using pundit gem as a pattern to structure the permission logic.

* Also it's required to run the `rake db:seed` when initializing the app
to setup the Resource support table that handles the list of available
resources( or record classes ) to be authorized to users.

* The permisisons table has 2 fields(resourceable_id, resourceable_type)
working as polymorphic fields so the user can add a permission to a
single record or just omit the resourceable_id field and add an ability
to all records to the referent model filled in the resourceable_type
field.

* Resource model is not added to any tenant because it's a table used by
the entire app as configuration data.

* All models that will have permissions have to have the has_many
relationship to permissions. The Role model needed a fix to work due the
conflict of having many permissions and having many specific record
permissions.

* Any test in rspec was made at this point referent to this
modifications.